pyproject.toml
[tool.poetry]
name = "capfinder"
version = "0.1.3"
description = "A package for decoding RNA cap types"
authors = [
    "Adnan M. Niazi <adnaniazi@gmail.com>",
]
license = "MIT"
readme = "README.md"

documentation = "https://adnaniazi.github.io/capfinder"
homepage = "https://adnaniazi.github.io/capfinder"
repository = "https://github.com/adnaniazi/capfinder"

classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "Operating System :: OS Independent",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.9",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "Typing :: Typed",
]

packages = [
    { include = "capfinder", from = "src" }
]

[tool.poetry.dependencies]
python = ">=3.9, <3.10"
click = "*"
pysam = "^0.21.0"
parasail = "^1.3.4"
mpire = "^2.8.0"
tqdm = "^4.66.1"
pytest-mock = "^3.11.1"
matplotlib = "^3.8.0"
loguru = "^0.7.2"
bokeh = "^3.3.0"
toml = "^0.10.2"
types-pkg-resources = "^0.1.3"
types-toml = "^0.10.8.7"
types-tqdm = "^4.66.0.20240106"
keras = "^3.1.1"
pod5 = "^0.3.10"
comet-ml = "^3.39.3"
prefect = "^2.17.1"
prefect-dask = "^0.2.6"
polars = "^0.20.21"
tensorflow = "^2.16.1"
hpack = "^4.0.0"
httpx = "^0.27.0"
biopython = "^1.83"
pre-commit = "^3.7.0"


[tool.poetry.dev-dependencies]
autoflake = "*"
black = "*"
flake8 = "*"
flake8-bugbear = "*"
flake8-builtins = "*"
flake8-comprehensions = "*"
flake8-debugger = "*"
flake8-eradicate = "*"
flake8-logging-format = "*"
isort = "*"
mkdocstrings = {version = ">=0.18", extras = ["python"]}
mkdocs-material = "*"
mypy = "*"
pep8-naming = "*"
pre-commit = "*"
pymdown-extensions = "*"
pytest = "*"
pytest-github-actions-annotate-failures = "*"
pytest-cov = "*"
python-kacl = "*"
pyupgrade = "*"
tryceratops = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
src_paths = ["src", "tests"]

[tool.black]
target-version = ["py37", "py38", "py39"]
include = '\.pyi?$'

[tool.pytest.ini_options]
addopts = """\
    --cov capfinder \
    --cov tests \
    --cov-report term-missing \
    --no-cov-on-fail \
"""

[tool.coverage.report]
fail_under = 0
exclude_lines = [
    'if TYPE_CHECKING:',
    'pragma: no cover'
]

[tool.mypy]
disallow_any_unimported = true
disallow_untyped_defs = true
no_implicit_optional = true
strict_equality = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_return_any = true
check_untyped_defs = true
show_error_codes = true
ignore_missing_imports = true

#[tool.poetry.scripts]
#add = "capfinder.add:add"

#command1 = "package:module1:function1"
#command2 = "package:module2:function2"
#command3 = "package:module3:function3"

[tool.setuptools_scm]
